+++ src/lib/images/index.ts
@@ -1,6 +1,11 @@
 export {default as icon} from './icon.svg';
 export {default as logo} from './logo.svg';
 export {default as logoBeside} from './logo-beside.svg';
+
 export {default as github} from './github.svg';
 export {default as session} from './session.svg';
 export {default as tor} from './tor.svg';
+
+export {default as info} from './info.svg';
+export {default as alert} from './alert.svg';
+export {default as success} from './success.svg';