+++ svelte.config.js
@@ -1,4 +1,3 @@
-// import adapter from '@sveltejs/adapter-bun';
 import adapter from '@sveltejs/adapter-auto';
 import {vitePreprocess} from '@sveltejs/vite-plugin-svelte';