+++ CHANGELOG.md
@@ -5,6 +5,25 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.0.4] - 2024-10-03
+
+### Added
+
+- Add Notification component
+- Add info/success/error icons
+- Add svelte alias for ease of use
+- Add stores.ts for stores
+- Add types.ts for types
+
+### Changed
+
+- Destructure index.ts into multiple files
+- Refecture code
+
+### Removed
+
+- Remove svelte contextAPI in favor of stores
+
 ## [0.0.3] - 2024-10-02
 
 ### Added